manifest: s/configuration/parameters
The manifest used to contain an entry called "configuration" that was
a Hash of name to Parameter options. That made it impossible to order
them. This changes the key to "parameters" (to match what they are)
and changes the value to an Array of Hashes, each with a "name"
field.